anpr_system/requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pytesseract>=0.3.10
firebase-admin>=6.2.0
python-dotenv>=1.0.0
pytest>=7.4.0
pillow>=10.0.0
imutils>=0.5.4 